Issue 8: ElasticCacheClient not always dynamically refreshing
If the first expansion turns out to fail or return a single
node, DynamicResolvingClient would conclude that the host list
was a static host list, which is never the case for subclasses
that override expand_entry with dynamic logic.

So, add a class attribute to mark always-dynamic subclasses
as such, and make ElasticCacheStoreClient such a subclass.