Implement direct passthroughs for get
Wrapper clients (decorated, namespace, async) need direct
passthroughs for get/getMulti (without TTL) to not deny the
optimizations that don't load/unpickle values when ttl_skip
excludes a response. Otherwise the wrappers will turn the
get into a getTtl, that will have to load the value even if
stale to report it.