Clear event when stuck busy looping
It can happen that noone clears done_event, especially when
the pool is configured with non-blocking behavior. So if
waitkey finds itself busy looping on a done_event that doesn't
get cleared, make it clear it to go back to normal.

This could cause a latency spike in enqueue if the event is
cleared before it starts to wait on it, but it seems unlikely,
and if it does happen, it would only mean a single 1s latency
spike.