Further reduce latency
By using timeout-less waiting in WaitIter when threadpools
are waiting for new tasks. Idle threads will otherwise incur
the hardcoded minimum delay of 500us, or if they've been idle
for a while, maybe even the maximum hardcoded delay of 50ms,
which can be a lot in some contexts.

Waiting on an event without timeout has no problem, we just
have to be extra careful to ensure threads do get eventually
woken up.