Reorder operations to prevent race conditions with GC
It's unlikely and unlikely to cause issues, but this reordering
makes it moot. Reorder operations to avoid race conditions when
list copying invokes the GC. If that causes re-entrancy due to
some other thread waking up and invoking set, the original thread
would fail.

In this new code, that can't happen.