Optimize code generation for native platform
Unless we build a wheel, in which case we can override with environment
variables, the platform building will be the platform running as
pip will install from source.

In that case, passing `-mtune=native` to GCC can speed up code, and it's
a very lightweight change so it's more than worth the trouble.