PY3: Change pickler interface
To make more sense, and less error prone. Make it a regular
strategy object instead of a weird namespace thing that is
hard to configure.

Also make it encapsulate all the separators and protocol
decisions to remove that concern from the memcache client itself,
simplifying things when non-pickle or non-json picklers
are used.